// src/app/actiontracker/model/actiontrackerFilter.model.ts
export class ActiontrackerFilter {
    UserId: number;
    RoleId: number;
    RoleName: string;
    OrderdByColoumName: string;
    OrderdBy: string;
    ProcessIds: number[];
    PriorityTypes: number[];
    StatusId: number[];
    PageSize: number;
    PageIndex: number;
    DashboardFor: string;
    ActionAssignFromDate: Date;
    ActionAssignToDate: Date;
}